requirements.txt
streamlit
python-dotenv
google-genai
exa_py
pypdf
langchain
langchain-chroma
langchain-google-genai
langchain-community